Microservice_Backend API/app/requirements.txt
pydantic==1.8.2
fastapi==0.67.0
tensorflow
pillow
numpy
requests

#Run this file with - pip3 install -r requirements.txt
